backend/models/throat_cancer/requirements.txt
torch==2.6.0
torchvision==0.21.0
numpy==1.24.3
pillow==10.0.0
scikit-learn==1.3.0
